Bug fix in primes python

+++ primes/primes.py
@@ -38,7 +38,7 @@ if (low % 2 == 0):
 for i in range(low, high, 2):
   flag = 0
 
-  for j in range(2, i/2):
+  for j in range(2, int(i/2)):
     if (i % j == 0):
       flag = 1
       break
